requirements.txt
Flask==2.3.2
python-dotenv==1.0.0
Flask-PyMongo==2.3.0
Flask-Bcrypt==1.0.1
PyJWT==2.8.0
